- Added mouse rendering so that the mouse only shows up in game and the cursor changes based on context (boosting vs. not boosting). - Updated star milestone locks so that they're more in line with what a average player is capable of (starting around 70-75% and works up to 80-81% from start of classical to the end of deco) - Small visual polish tweaks

+++ trunk/GameModel/BallInPlayState.cpp
@@ -116,7 +116,7 @@ void BallInPlayState::MoveKeyPressedForOther(int dir, float magnitudePercent) {
     }
 }
 
-void BallInPlayState::BallBoostDirectionPressed(int x, int y) {
+void BallInPlayState::BallBoostDirectionPressed(float x, float y) {
     assert(this->gameModel->boostModel != NULL);
 
     // Ignore applying the boost direction when the ball is paused
